requirements.txt
# Biofloc Monitor UCN - Dependencies
# Python 3.10+ required

# Web Framework
streamlit>=1.36.0

# Data Processing
pandas>=2.0.0
numpy>=1.26.0

# Database
pymongo>=4.0.0
certifi>=2024.0.0

# Environment Variables
python-dotenv>=1.0.0

# Visualization
plotly>=5.18.0

# Excel Export
openpyxl>=3.1.0